new: authentication added

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const contactModel = require('../models/contactModel')
 
 // @desc Get all contacts
 // @route GET /api/contacts
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res) => {
-    const contacts = await contactModel.find()
+    const contacts = await contactModel.find({ user_id: req.user.id })
     res.status(200).json({ contacts })
 })
 
 // @desc Get Single contacts
 // @route GET /api/contacts/:id
-// @access public
+// @access private
 const getSingleContact = asyncHandler(async (req, res) => {
     const contact = await contactModel.findById(req.params.id)
     if (!contact) {
@@ -23,7 +23,7 @@ const getSingleContact = asyncHandler(async (req, res) => {
 
 // @desc Create new contacts
 // @route POST /api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
     const { name, email, phone } = req.body
     if (!name || !email || !phone) {
@@ -31,35 +31,44 @@ const createContact = asyncHandler(async (req, res) => {
         throw new Error('All fields are required')
     }
 
-    const contacts = await contactModel.create({ name, email, phone })
-    console.log(req.body)
+    const contacts = await contactModel.create({ name, email, phone, user_id: req.user.id })
 
     res.status(200).json(contacts)
 })
 
 // @desc Update contacts
 // @route PUT /api/contacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await contactModel.findById(req.params.id)
     if (!contact) {
         res.status(404)
         throw new Error('Contact not found')
     }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403)
+        throw new Error("User dont have permission")
+    }
+
     const updatedContact = await contactModel.findByIdAndUpdate(req.params.id, req.body, { new: true })
     res.status(200).json(updatedContact)
 })
 
 // @desc Delete contacts
 // @route DELETE /api/contacts/:id
-// @access public
+// @access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await contactModel.findById(req.params.id)
     if (!contact) {
         res.status(404)
         throw new Error('Contact not found')
     }
-    await contact.deleteOne({_id:req.params.id})
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403)
+        throw new Error("User dont have permission")
+    }
+
+    await contact.deleteOne({ _id: req.params.id })
     res.status(200).json(contact)
 })
